implement timer for futures
This is a new crate which provides a timer for futures. It wraps a
future and once it completes, includes timing information as the output.
It records the busy and idle time of the inner future.

The crate is mostly complete, but needs some reword to type erase the
`FutureTimer`, which will is more forward compatible, especially if we
want to provide a `no_std` version in the future which swaps out
`Instant` for something provided by the user.