chore: set up CI builds (#1)
Add a CI workflow for GitHub actions.

It runs check, tests on multiple OSes and also runs cargo fmt and
clippy.

This change includes some formatting fixes and also replaces the
`Cargo.lock` file with one using version = 3 of the lock file format so
that Rust 1.70 can be supported. This is the oldest version we can
support and also support `mio` which is used (via `tokio`) in out tests.