chore: rename crate to future-timing
We're interested in timing a future, and a timer may be considered more
something like a future that resolves after a specific time.

A number of types have also been renamed to simplify.

Building the docs has been added to the CI and the lints we're using
have been specified in the cargo manifest.